tools: better test.dart
