--- tools/test_result.dart
@@ -1,11 +1,0 @@
-
-class TestResult {
-  final String module;
-  final int exitCode;
-  bool get passed => exitCode == 0;
-
-  TestResult(this.module, this.exitCode);
-  
-  @override
-  String toString() => '$module tests ${exitCode == 0 ? 'passed' : 'FAILED'}';
-}
